Move panel files to folder and add snapshot tests
